separate solver files and adjust all includes
